Use `unsigned short int` for serialization sample.

+++ tests/engine_lib/io/Serializable.cpp
@@ -752,7 +752,7 @@ TEST_CASE("serialize and deserialize sample player save data") {
         // Prepare new profile file name.
         std::random_device rd;
         std::mt19937 gen(rd());
-        std::uniform_int_distribution<unsigned int> uid(0, UINT_MAX);
+        std::uniform_int_distribution<unsigned short int> uid(0, USHRT_MAX);
         std::string sNewProfileFilename;
         const auto vExistingProfiles = ConfigManager::getAllFiles(ConfigCategory::PROGRESS);
         bool bContinue = false;
